REFRENCES/jarvis-adk-voice-agent/requirements.txt
google-adk
fastapi
uvicorn[standard]
websockets
python-dotenv
certifi
PyPDF2
litellm
